requirements.txt
langchain>=0.1.0
streamlit>=1.40.1
faiss-cpu>=1.7.4
openai>=1.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.3